tools/names/scraper_requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
psycopg2-binary>=2.9.0
lxml>=4.9.0